Fix root cause: get_team_roster() bigint/numeric type mismatch
team_roster_view.total_xp resolves to bigint (sum(int) over
xp_ledger.amount), but get_team_roster() declared it numeric in its
RETURNS TABLE — a genuine structural mismatch Postgres rejects. This
was silently breaking roster/team-XP loading since Phase C: it's one
of three calls bundled in a Promise.all with no error handling, so
its failure quietly prevented the other two (which worked fine) from
ever being applied, showing "0" everywhere with no visible cause.
Found via the on-page error message added last commit. Cast to
::numeric in the function's return query.

Also removes the temporary [debug] status text now that the real
cause is confirmed — the timeout/error-surfacing wrapper stays.

+++ supabase/migrations/0003_views.sql
@@ -72,7 +72,7 @@ begin
   ) then
     raise exception 'not authorized for this team roster';
   end if;
-  return query select v.team_id, v.athlete_id, v.display_name, v.total_xp, v.workout_count, v.last_workout_date, v.status
+  return query select v.team_id, v.athlete_id, v.display_name, v.total_xp::numeric, v.workout_count, v.last_workout_date, v.status
     from team_roster_view v where v.team_id = p_team_id;
 end $$;
